Directional Trading
Momentum trading is trading when the instrument is trending up or down.  With my algorithm, I bet on the direction of the instrument -- I aim
for buying at a low price and selling at a high price. I show how to implement momentum trading rules and backtest using
historical GOOGLE stock data in R.